+++ src/issuingCard/issuingCard.php
@@ -21,9 +21,9 @@ class IssuingCard extends Resource
 
     ## Parameters (optional):
         - displayName [string, default null]: card displayed name. ex: "ANTHONY STARK"
-        - rules [list of IssuingRule, default null]: [EXPANDABLE] list of card spending rules.
+        - rules [array of IssuingRule, default null]: [EXPANDABLE] array of card spending rules.
         - binId [string, default null]: BIN ID to which the card is bound. ex: "53810200"
-        - tags [list of strings]: list of strings for tagging. ex: ["travel", "food"]
+        - tags [array of strings]: array of strings for tagging. ex: ["travel", "food"]
         - streetLine1 [string, default null]: card holder main address. ex: "Av. Paulista, 200"
         - streetLine2 [string, default null]: card holder address complement. ex: "Apto. 123"
         - district [string]: card holder address district / neighbourhood. ex: "Bela Vista"
@@ -32,15 +32,15 @@ class IssuingCard extends Resource
         - zipCode [string, null]: card holder address zip code. ex: "01311-200"
 
     ## Attributes (return-only):
-        - id [string, default null]: unique id returned when IssuingCard is created. ex: "5656565656565656"
-        - holderId [string, default null]: card holder unique id. ex: "5656565656565656"
-        - type [string, default null]: card type. ex: "virtual"
-        - status [string, default null]: current IssuingCard status. ex: "canceled" or "active"
-        - number [string, default null]: [EXPANDABLE] masked card number. ex: "1234 5678 1234 5678"
-        - securityCode [string, default null]: [EXPANDABLE] masked card verification value (cvv). Expand to unmask the value. ex: "123".
-        - expiration [string, default null]: [EXPANDABLE] masked card expiration datetime. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
-        - created [string, default null]: creation datetime for the IssuingCard. ex: "2020-03-10 10:30:00.000"
-        - updated [string, default null]: latest update datetime for the IssuingCard. ex: "2020-03-10 10:30:00.000"
+        - id [string]: unique id returned when IssuingCard is created. ex: "5656565656565656"
+        - holderId [string]: card holder unique id. ex: "5656565656565656"
+        - type [string]: card type. ex: "virtual"
+        - status [string]: current IssuingCard status. ex: "canceled" or "active"
+        - number [string]: [EXPANDABLE] masked card number. ex: "1234 5678 1234 5678"
+        - securityCode [string]: [EXPANDABLE] masked card verification value (cvv). Expand to unmask the value. ex: "123".
+        - expiration [string]: [EXPANDABLE] masked card expiration datetime. 
+        - created [DateTime]: creation datetime for the IssuingCard. 
+        - updated [DateTime]: latest update datetime for the IssuingCard. 
      */
     function __construct(array $params)
     {
@@ -74,22 +74,22 @@ class IssuingCard extends Resource
     /**
     # Create IssuingCards
 
-    Send a list of IssuingCard objects for creation in the Stark Infra API
+    Send an array of IssuingCard objects for creation in the Stark Infra API
 
     ## Parameters (required):
-        - cards [list of IssuingCard objects]: list of IssuingCard objects to be created in the API
+        - cards [array of IssuingCard objects]: array of IssuingCard objects to be created in the API
 
     ## Parameters (optional):
-        - expand [list of strings, default []]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
+        - params [dictionary of optional parameters]:
+            - expand [array of strings, default null]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
-        - list of IssuingCard objects with updated attributes
+        - array of IssuingCard objects with updated attributes
      */
-    public static function create($cards, $expand = null, $user = null)
+    public static function create($cards, $params = null, $user = null)
     {
-        $query = is_null($expand) ? [] : ["expand" => $expand];
-        return Rest::post($user, IssuingCard::resource(), $cards, $query);
+        return Rest::post($user, IssuingCard::resource(), $cards, $params);
     }
 
     /**
@@ -100,13 +100,13 @@ class IssuingCard extends Resource
     ## Parameters (optional):
         - status [string, default null]: filter for status of retrieved objects. ex: "paid" or "registered"
         - types [string, default null]: card type. ex: "virtual"
-        - holderIds [list of strings]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
-        - after [datetime.date or string, default null] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
-        - before [datetime.date or string, default null] date filter for objects created only before specified date. ex: datetime.date(2020, 3, 10)
-        - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
-        - ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - holderIds [array of strings]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
+        - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"
+        - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
+        - tags [array of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+        - ids [array of strings, default null]: array of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
-        - expand [list of strings, default []]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
+        - expand [array of strings, default []]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
@@ -123,27 +123,30 @@ class IssuingCard extends Resource
     /**
     # Retrieve paged IssuingCards
 
-    Receive a list of IssuingCards objects previously created in the Stark Infra API and the cursor to the next page.
+    Receive an array of IssuingCards objects previously created in the Stark Infra API and the cursor to the next page.
 
     ## Parameters (optional):
         - cursor [string, default null]: cursor returned on the previous page function call
         - status [string, default null]: filter for status of retrieved objects. ex: "paid" or "registered"
         - types [string, default null]: card type. ex: "virtual"
-        - holderIds [list of strings]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
-        - after [datetime.date or string, default null] date filter for objects created only after specified date. ex: datetime.date(2020, 3, 10)
-        - before [datetime.date or string, default null] date filter for objects created only before specified date. ex: datetime.date(2020, 3, 10)
-        - tags [list of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
-        - ids [list of strings, default null]: list of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
+        - holderIds [array of strings]: card holder IDs. ex: ["5656565656565656", "4545454545454545"]
+        - after [Date or string, default null] date filter for objects created only after specified date. ex: "2020-04-03"  
+        - before [Date or string, default null] date filter for objects created only before specified date. ex: "2020-04-03"
+        - tags [array of strings, default null]: tags to filter retrieved objects. ex: ["tony", "stark"]
+        - ids [array of strings, default null]: array of ids to filter retrieved objects. ex: ["5656565656565656", "4545454545454545"]
         - limit [integer, default null]: maximum number of objects to be retrieved. Unlimited if null. ex: 35
-        - expand [list of strings, default []]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
+        - expand [array of strings, default []]: fields to to expand information. ex: ["rules", "securityCode", "number", "expiration"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
     
     ## Return:
-        - list of IssuingCards objects with updated attributes
+        - array of IssuingCards objects with updated attributes
         - cursor to retrieve the next page of IssuingCards objects
      */
     public static function page($options = [], $user = null)
     {
+        $options["after"] = new StarkDate(Checks::checkParam($options, "after"));
+        $options["before"] = new StarkDate(Checks::checkParam($options, "before"));
+
         return Rest::getPage($user, IssuingCard::resource(), $options);
     }
 
@@ -156,14 +159,16 @@ class IssuingCard extends Resource
         - id [string]: object unique id. ex: "5656565656565656"
 
     ## Parameters (optional):
+        - params [dictionary of optional parameters]:
+            - expand [array of strings, default null]: fields to expand information. ex: ["rules", "security_code", "number", "expiration"]
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
         - IssuingCards object with updated attributes
      */
-    public static function get($id, $user = null)
+    public static function get($id, $params=null, $user = null)
     {
-        return Rest::getId($user, IssuingCard::resource(), $id);
+        return Rest::getId($user, IssuingCard::resource(), $id, $params);
     }
 
     /**
@@ -177,8 +182,8 @@ class IssuingCard extends Resource
     ## Parameters (optional):
         - status [string]: You may block the IssuingCard by passing 'blocked' in the status
         - displayName [string, default null]: card displayed name
-        - rules [list of dictionaries, default null]: list of dictionaries with "amount": int, "currencyCode": string, "id": string, "interval": string, "name": string pairs.
-        - tags [list of strings]: list of strings for tagging
+        - rules [array of dictionaries, default null]: array of dictionaries with "amount": int, "currencyCode": string, "id": string, "interval": string, "name": string pairs.
+        - tags [array of strings]: array of strings for tagging
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
 
     ## Return:
@@ -189,11 +194,10 @@ class IssuingCard extends Resource
         return Rest::patchId($user, IssuingCard::resource(), $id, $options);
     }
 
-
     /**
-    # Delete a IssuingCard entity
+    # Cancel an IssuingCard entity
 
-    Delete a IssuingCard entity previously created in the Stark Infra API
+    Cancel an IssuingCard entity previously created in the Stark Infra API
 
     ## Parameters (required):
         - id [string]: IssuingCard unique id. ex: "5656565656565656"
@@ -202,9 +206,9 @@ class IssuingCard extends Resource
         - user [Organization/Project object, default null]: Organization or Project object. Not necessary if StarkInfra\Settings::setUser() was used before function call
     
     ## Return:
-        - deleted IssuingCard object
+        - canceled IssuingCard object
      */
-    public static function delete($id, $user = null)
+    public static function cancel($id, $user = null)
     {
         return Rest::deleteId($user, IssuingCard::resource(), $id);
     }